Break skiddle API calls out into separate module
