+++ controllers/events.js
@@ -2,12 +2,7 @@ var express = require('express'),
     router = express.Router();
 var request = require('request');
 var fs = require('fs');
-
-// API keys to access the skiddle festival database
-// var contents = fs.readFileSync(__dirname + '/../config/api_keys.json');
-// var api_keys = JSON.parse(contents);
-
-var cachedEventData = null;
+var skiddleAPI = require(__dirname + '/../utils/skiddleAPI');
 
 // Serve requests to the event endpoint
 router.get('/event', function(req, res){
@@ -21,37 +16,7 @@ router.get('/event', function(req, res){
         // 'All' returns the entire events list
         if(req.query.type === "all") {
 
-            // Check to see if we've already cached the event data
-            if (!cachedEventData) {
-                //Get all events from Skiddle API
-                var options = { method: 'GET',
-                  url: process.env.skiddle_url  + 'events/',
-                  qs: 
-                   { eventcode: 'FEST',
-                     order: '4',
-                     api_key: process.env.skiddle_api_key } };
-
-                console.log("GET " + options.url);
-
-                //Send the request
-                request(options, function (error, reqResponse, body) {
-                  if (error) throw new Error(error);
-                  var contents = JSON.parse(body);
-                  if (contents.error !== 0) {
-                    console.log('ERROR ' + contents.errorcode + ': ' + contents.errormessage);
-                  } else {
-                      cachedEventData = contents.results;
-                      response = contents.results;
-                      response.ok = true;
-                  }
-                  res.send(response);
-                });
-            // Cached data present
-            } else {
-                response = cachedEventData;
-                response.ok = true;
-                res.send(response);
-            }
+            skiddleAPI.getAllEvents(res);
 
         // Unknown query type
         } else {
